Configure all ips on one interface
Instead of attaching one interface per BMC address, this change
assigns exactly two interfaces to the BMC, using one via DHCP for
the default route and as a basic management IP in case of issues.
The second is assigned all of the BMC addresses allocated by the
Heat template.

I'm trying this method for a few reasons:

-Having multiple interfaces seems to cause networking issues.
 Even with only five baremetal instances, the bmc networking was
 flaky enough that I couldn't do a deployment.  Using only two
 interfaces does still have two routes to the private subnet,
 but it doesn't seem to cause problems in my testing so far.  The
 jury's still out on whether this is a solution or simply mitigates
 the problem by limiting the number of duplicate routes.

 For reference, while I'm not 100% sure what was wrong here, it
 seemed like maybe there was something weird going on with the
 UDP protocol and the multiple routes to the private subnet. I
 was seeing errors in tcpdump where the BMC was unable to talk
 back to the UDP port on the undercloud when I tried to
 communicate with the bmc service.
-Adding a large number of Neutron ports on a single VM seems to
 trigger https://bugs.launchpad.net/neutron/+bug/1515768
 While this bug should also be fixed, it is currently a blocker for
 any sort of scale testing with the previous single bmc method.
-Along similar lines, this method is less likely to run up against
 limitations of Neutron/Nova/Libvirt/all the things involved in
 attaching an interface to an instance. All it's doing is assigning
 additional IPs to a single interface using os-net-config, which
 should only be limited by the OS's ability to handle large numbers
 of IP addresses.
-It will make the Horizon instances view less ugly when a large
 number of instances are attached to one bmc.  This method only has
 two IPs displayed, whereas the many interfaces method would have
 as many IPs as there are instances.  Imagine Horizon with a
 100 node baremetal stack. :-)

One thing to note is that currently the implementation wastes a
private IP.  Since this is a private neutron network where IPs
should be essentially "free", I don't think this is a significant
problem, but it's possible we could eliminate the wasted IP by
assigning that port the first BMC port instead of letting Neutron
generate a new one.  I didn't know off the top of my head how to
do that though.

Also, while it would be possible to do this with a single port on
the BMC instance, that would require us to set the default route
manually with os-net-config, and it means that if anything went
wrong with the os-net-config run there would be no way to get into
the instance.  Two interfaces avoids both those problems.