Remove controller config from templates
Upon further reflection, I think it's more flexible and maintainable
to simply do this using a separate stack.  If the prefixes are
chosen appropriately (say baremetal and baremetal_controller), the
build-nodes-json script will include all of the nodes in the json
file, so we end up with essentially the same thing, but you can
reasonably do this for an arbitrary number of different nodes types.

This also doesn't play nicely with the single BMC changes, so that's
another motivation for making the change.  We could fix the single
BMC case, but as noted above I prefer the multi-stack method
anyway.

For convenience, I added a deploy-controller script that takes its
environment configuration from a controller.yaml file instead of
env.yaml.